requirements.txt
Flask==2.0.1
flask-cors
supabase
python-dotenv
werkzeug
gunicorn
